[detail_item] 1.코드리뷰 반영 및 수정완료
