+++ src/Components/DetailItem/DetailItem.scss
@@ -12,7 +12,7 @@
     @include flexBox;
     width: 1300px;
     padding-top: 70px;
-    margin: 0 310px 0 310px;
+    margin: 0 310px;
 
     .itemBox {
       display: flex;
