+++ src/Components/DetailItem/DropDownBox.js
@@ -1,7 +1,8 @@
 import React, { Component } from 'react';
 import Select from 'react-select';
-import './DropDownBox.scss';
 import OrderBox from './OrderBox';
+import { COLOR, SIZE } from './ITEM';
+import './DropDownBox.scss';
 
 class DropDownBox extends Component {
   constructor(props) {
@@ -113,8 +114,8 @@ class DropDownBox extends Component {
           <Select
             className="select1"
             onChange={this.handleChange}
-            options={COLLOR}
-            defaultValue={COLLOR[0]}
+            options={COLOR}
+            defaultValue={COLOR[0]}
           />
           {number > 0 && (
             <Select
@@ -171,17 +172,3 @@ class DropDownBox extends Component {
 }
 
 export default DropDownBox;
-
-const COLLOR = [
-  { id: 0, value: 'default', label: '[컬러]를 선택하세요.' },
-  { id: 1, value: 'Black', label: '블랙' },
-  { id: 2, value: 'Yellow', label: '옐로우' },
-  { id: 3, value: 'Gray', label: '그레이' },
-];
-
-const SIZE = [
-  { id: 0, value: 'default', label: '[사이즈]를 선택하세요.' },
-  { id: 1, value: 'Small', label: 'Small' },
-  { id: 2, value: 'Medium', label: 'Medium' },
-  { id: 3, value: 'Large', label: 'Large' },
-];
