+++ src/Components/DetailItem/ITEM.js
@@ -1,0 +1,13 @@
+export const COLOR = [
+  { id: 0, value: 'default', label: '[컬러]를 선택하세요.' },
+  { id: 1, value: 'Black', label: '블랙' },
+  { id: 2, value: 'Yellow', label: '옐로우' },
+  { id: 3, value: 'Gray', label: '그레이' },
+];
+
+export const SIZE = [
+  { id: 0, value: 'default', label: '[사이즈]를 선택하세요.' },
+  { id: 1, value: 'Small', label: 'Small' },
+  { id: 2, value: 'Medium', label: 'Medium' },
+  { id: 3, value: 'Large', label: 'Large' },
+];
