+++ src/Components/DetailItem/STAR.js
@@ -1,0 +1,9 @@
+const STAR = [
+  { id: 1, img: './images/star.png' },
+  { id: 2, img: './images/star.png' },
+  { id: 3, img: './images/star.png' },
+  { id: 4, img: './images/star.png' },
+  { id: 5, img: './images/star.png' },
+];
+
+export default STAR;
