+++ src/common.scss
@@ -1,3 +1,9 @@
 * {
   background-color: #ffffff;
 }
+
+@mixin flexBox {
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
